+++ internal/app/server/usecase.go
@@ -4,11 +4,11 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/lidofinance/finding-forwarder/internal/app/feeder"
-	"github.com/lidofinance/finding-forwarder/internal/connectors/metrics"
-	"github.com/lidofinance/finding-forwarder/internal/env"
-	"github.com/lidofinance/finding-forwarder/internal/pkg/chain"
-	"github.com/lidofinance/finding-forwarder/internal/pkg/notifiler"
+	"github.com/lidofinance/onchain-mon/internal/app/feeder"
+	"github.com/lidofinance/onchain-mon/internal/connectors/metrics"
+	"github.com/lidofinance/onchain-mon/internal/env"
+	"github.com/lidofinance/onchain-mon/internal/pkg/chain"
+	"github.com/lidofinance/onchain-mon/internal/pkg/notifiler"
 )
 
 type Services struct {